Python/Sync: Move protobuf folder into the shared folder
