+++ python/DEVELOPER.md
@@ -195,7 +195,7 @@ pytest -v --async-backend=trio --async-backend=asyncio
 
 # Protobuf
 ---
-During the initial build, Python protobuf files were created in `python/python/glide/protobuf`. If modifications are made to the protobuf definition files (`.proto` files located in `glide-core/src/protofuf`), it becomes necessary to regenerate the Python protobuf files.
+During the initial build, Python protobuf files were created in `python/python/glide/shared/protobuf`. If modifications are made to the protobuf definition files (`.proto` files located in `glide-core/src/protofuf`), it becomes necessary to regenerate the Python protobuf files.
 
 Protobuf files are automatically regenerated as part of the `build` process. If you only need to regenerate the protobuf files (e.g., after editing `.proto` files), you can use the standalone command the root `python/` directory:
 
@@ -203,7 +203,7 @@ Protobuf files are automatically regenerated as part of the `build` process. If 
 python3 dev.py protobuf
 ```
 
-This generates `.py` and `.pyi` interface files for type checking and places them in the `python/python/glide/protobuf` folder.
+This generates `.py` and `.pyi` interface files for type checking and places them in the `python/python/glide/shared/protobuf` folder.
 
 # Linters
 ---
