+++ python/docs/conf.py
@@ -28,7 +28,7 @@ templates_path = ["_templates"]
 exclude_patterns = ["_build", "Thumbs.db", ".DS_Store"]
 autodoc_mock_imports = [
     "glide.glide",
-    "glide.protobuf",
+    "glide.shared.protobuf",
     "pytest",
     "google",
     "cffi",
