+++ python/tests/test_config.py
@@ -1,8 +1,5 @@
 # Copyright Valkey GLIDE Project Contributors - SPDX Identifier: Apache-2.0
 
-from glide.protobuf.connection_request_pb2 import ConnectionRequest
-from glide.protobuf.connection_request_pb2 import ReadFrom as ProtobufReadFrom
-from glide.protobuf.connection_request_pb2 import TlsMode
 from glide.shared.config import (
     AdvancedGlideClientConfiguration,
     AdvancedGlideClusterClientConfiguration,
@@ -15,6 +12,9 @@ from glide.shared.config import (
     PeriodicChecksStatus,
     ReadFrom,
 )
+from glide.shared.protobuf.connection_request_pb2 import ConnectionRequest
+from glide.shared.protobuf.connection_request_pb2 import ReadFrom as ProtobufReadFrom
+from glide.shared.protobuf.connection_request_pb2 import TlsMode
 
 
 def test_default_client_config():
